pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pauline-neural-study"
version = "0.1.0"
description = "A novel computational approach to analyzing the Pauline corpus using bootstrap resampling, combinatorial recombination, and neural embeddings — operating exclusively on Paul's own words."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Pauline Neural Study Contributors"},
]
keywords = [
    "pauline-epistles",
    "biblical-studies",
    "computational-linguistics",
    "word-embeddings",
    "bootstrap-resampling",
    "digital-humanities",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "gensim>=4.3.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
viz = ["umap-learn>=0.5.0"]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[project.scripts]
pauline-study = "pauline.pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
